refactor: add constants, move some dnd methods to dndUtils.
