--- src/utils/roundToGrid.js
@@ -1,4 +1,0 @@
-export default function roundToGrid(value, gridSize) {
-  const rounded = Math.round(value / gridSize);
-  return parseFloat((rounded * gridSize).toFixed(2));
-}
